+++ src/templates/react-app/db/indexedDb/loop/{{modelNameCamel}}.mustache.js
@@ -1,0 +1,6 @@
+import { withStore } from './db';
+const { add, get, set, deleteRecord, getAll } = withStore(
+  '{{modelNamePascal}}'
+);
+
+export { get, add, set, deleteRecord, getAll };